# doc/openapi.yaml
openapi: 3.0.3
info:
  title: 微信群搭子导流平台 API
  description: |
    基于 PRD v1.3 / 技术方案 v1.0。
    - Base URL: `/api/v1`
    - 统一响应: `{ code, message, data, requestId }`
    - C 端鉴权: `Authorization: Bearer <jwt>`
    - 管理后台: `Authorization: Bearer <admin_jwt>`
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000/api/v1
    description: 本地开发
  - url: https://www.example.com/api/v1
    description: 生产环境
...
paths:
...
  /groups:
    get:
      tags: [Group]
      summary: 群列表
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: regionId
          in: query
          schema:
            type: integer
        - name: hot
          in: query
          schema:
            type: boolean
          description: true=热门群 Tab
        - name: tagId
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/PaginatedGroups'
  /groups/{id}:
    get:
      tags: [Group]
      summary: 群详情（不含真实二维码）
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/GroupDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{id}/qrcode:
    get:
      tags: [Group]
      summary: 获取群二维码（需登录且已付费）
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: 返回短时效签名 URL
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/QrcodeUrl'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: 未付费
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: 40301
                message: PAYMENT_REQUIRED
                data: null
...
  /payments/wechat/prepay:
    post:
      tags: [Payment]
      summary: 发起微信支付
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderNo, scene]
              properties:
                orderNo:
                  type: string
                scene:
                  type: string
                  enum: [jsapi, h5]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/WechatPrepay'
  /payments/wechat/notify:
    post:
      tags: [Payment]
      summary: 微信支付回调（微信服务器调用）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 处理成功
...
  /admin/auth/login:
    post:
      tags: [Admin-Auth]
      summary: 管理员登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/AdminAuthToken'
  /admin/dashboard:
    get:
      tags: [Admin-Dashboard]
      summary: 仪表盘数据
      security:
        - AdminBearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /admin/groups:
    get:
      tags: [Admin-Group]
      summary: 群列表（管理）
      security:
        - AdminBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [Admin-Group]
      summary: 创建群
      security:
        - AdminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminGroupInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    GroupId:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    Unauthorized:
      description: 未登录
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            code: 40101
            message: UNAUTHORIZED
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
        requestId:
          type: string
          format: uuid
...
    GroupListItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        coverUrl:
          type: string
          nullable: true
        regionId:
          type: integer
        regionName:
          type: string
        tags:
          type: array
          items:
            type: string
        memberCount:
          type: integer
          nullable: true
        status:
          type: integer
          description: 1正常 2已满 3失效
    GroupDetail:
      allOf:
        - $ref: '#/components/schemas/GroupListItem'
        - type: object
          properties:
            description:
              type: string
            qrcodeLocked:
              type: boolean
              description: true=未付费，不返回真实二维码
    PaginatedGroups:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/GroupListItem'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    QrcodeUrl:
      type: object
      properties:
        url:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
...
    AdminAuthToken:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
          enum: [super, ops, cs, finance]
...
    WechatPrepay:
      type: object
      properties:
        appId:
          type: string
        timeStamp:
          type: string
        nonceStr:
          type: string
        package:
          type: string
        signType:
          type: string
        paySign:
          type: string
        h5Url:
          type: string
          description: H5 支付跳转链接
...
    AdminGroupInput:
      type: object
      required: [name, regionId, qrcodePath]
      properties:
        name:
          type: string
        coverUrl:
          type: string
        regionId:
          type: integer
        description:
          type: string
        memberCount:
          type: integer
        qrcodePath:
          type: string
        status:
          type: integer
        isHot:
          type: boolean
        weight:
          type: integer
        tagIds:
          type: array
          items:
            type: integer